Revert "Feat: Items Create Delete Frotnend and Chat Function"
