+++ frontend/src/pages/Home.css
@@ -1,4 +1,4 @@
-* {
+﻿* {
   box-sizing: border-box;
 }
 
@@ -105,7 +105,7 @@
 }
 
 .home-item-what {
-  font-size: 2rem;
+  font-size: 2.2rem;
   font-weight: 700;
   color: black;
   margin: 0;
@@ -116,7 +116,7 @@
 }
 
 .home-item-description {
-  font-size: 1.2rem;
+  font-size: 1.4rem;
   color: black;
   line-height: 1.4;
   font-weight: 400;
@@ -129,7 +129,7 @@
 .home-item-where,
 .home-item-location,
 .home-item-when {
-  font-size: 1.1rem;
+  font-size: 1.3rem;
   color: black;
   margin: 0;
   line-height: 1.3;
@@ -150,7 +150,7 @@
 }
 
 .home-founder-label {
-  font-size: 1.2rem;
+  font-size: 1.4rem;
   color: black;
   margin: 0;
   font-weight: 500;
@@ -161,7 +161,7 @@
 }
 
 .home-founder-name {
-  font-size: 1.6rem;
+  font-size: 1.8rem;
   font-weight: 600;
   color: black;
   margin: 0;
@@ -178,7 +178,7 @@
   border: none;
   border-radius: 4px;
   font-weight: 600;
-  font-size: 1.2rem;
+  font-size: 1.4rem;
   cursor: pointer;
   transition: transform 0.2s ease, box-shadow 0.2s ease;
   white-space: nowrap;
@@ -213,18 +213,17 @@
   background-color: var(--color-background-dark);
 }
 
-/*
 .home-empty-chat {
   display: flex;
   align-items: center;
   justify-content: center;
-  flex: 1;
+  height: 100%;
   color: #999;
   font-size: 1rem;
   text-align: center;
   padding: 1rem;
 }
-*/
+
 .home-page.dark-mode .home-empty-chat {
   color: #666;
 }
@@ -247,9 +246,6 @@ padding: 1.5vh 1vw;
   border-bottom: 2px solid var(--color-primary);
   background-color: var(--color-background-light);
   transition: background-color 0.3s ease;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
 }
 
 .home-page.dark-mode .home-chat-header {
@@ -277,18 +273,6 @@ padding: 1.5vh 1vw;
   color: white;
 }
 
-.online-indicator {
-  width: 12px;
-  height: 12px;
-  background-color: #00d084;
-  border-radius: 50%;
-  display: inline-block;
-}
-
-.online-indicator.disconnected {
-  background-color: #e63946;
-}
-
 .home-chat-messages {
   flex: 1;
   overflow-y: auto;
@@ -296,106 +280,6 @@ padding: 1.5vh 1vw;
   display: flex;
   flex-direction: column;
   gap: 1rem;
-}
-
-.message {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 0.5rem;
-  animation: slideIn 0.3s ease-in-out;
-}
-
-@keyframes slideIn {
-  from {
-    opacity: 0;
-    transform: translateY(10px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-.message.sent {
-  align-items: flex-end;
-}
-
-.message.received {
-  align-items: flex-start;
-}
-
-.message-bubble {
-  max-width: 85%;
-  padding: 0.75rem 1rem;
-  border-radius: 12px;
-  word-wrap: break-word;
-  word-break: break-word;
-}
-
-.message.sent .message-bubble {
-  background-color: var(--color-secondary);
-  color: black;
-  border-bottom-right-radius: 4px;
-}
-
-.message.received .message-bubble {
-  background-color: #e8e8e8;
-  color: #333;
-  border-bottom-left-radius: 4px;
-}
-
-.home-page.dark-mode .message.received .message-bubble {
-  background-color: #444444;
-  color: #e0e0e0;
-}
-
-.message-sender {
-  font-size: 0.75rem;
-  color: #666;
-  margin-bottom: 0.25rem;
-  font-weight: 500;
-}
-
-.message.sent .message-sender {
-  color: #666;
-  text-align: right;
-}
-
-.message.received .message-sender {
-  color: #999;
-  text-align: left;
-}
-
-.message-content {
-  word-break: break-word;
-  overflow-wrap: break-word;
-}
-
-.message-time {
-  font-size: 0.7rem;
-  color: #999;
-  margin-top: 0.25rem;
-}
-
-.message.sent .message-time {
-  text-align: right;
-  color: #999;
-}
-
-.message.received .message-time {
-  text-align: left;
-  color: #bbb;
-}
-
-.empty-messages {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: #999;
-  font-size: 0.9rem;
-  text-align: center;
-  padding: 2rem;
 }
 
 .home-chat-input-area {
@@ -677,155 +561,4 @@ padding: 1.5vh 1vw;
     transition: transform 0.2s ease, box-shadow 0.2s ease;
     white-space: nowrap;
   }
-}
-
-/* Notification styles */
-.home-chat-notifications-list {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  padding: 1rem;
-  background-color: #f0f0f0;
-  border-radius: 8px;
-  max-height: 300px;
-  overflow-y: auto;
-  margin-bottom: 1rem;
-}
-
-.home-page.dark-mode .home-chat-notifications-list {
-  background-color: #444;
-}
-
-/* Conversation list styles */
-.home-chat-conversations-list {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-  padding: 2vh;
-  padding-top: 6vh;
-  background-color: #f9f9f9;
-  border-radius: 0;
-  overflow-y: auto;
-  flex: 1;
-  height: 100%;
-}
-
-.home-page.dark-mode .home-chat-conversations-list {
-  background-color: #333;
-}
-
-.home-chat-conversations-list h3,
-.home-chat-notifications-list h3 {
-  margin: 0 0 0.5rem 0;
-  color: var(--color-primary);
-  font-size: 0.9rem;
-  text-transform: uppercase;
-}
-
-.conversation-item {
-  padding: 1vh;
-  background-color: white;
-  border: 1px solid #e0e0e0;
-  border-radius: 4px;
-  cursor: pointer;
-  transition: background-color 0.2s ease, transform 0.2s ease, border-color 0.2s ease;
-  display: flex;
-  flex-direction: column;
-  gap: 0.5vh;
-}
-
-.home-page.dark-mode .conversation-item {
-  background-color: #444;
-  border-color: #555;
-}
-
-.conversation-item:hover {
-  background-color: #f5f5f5;
-  border-color: var(--color-primary);
-  transform: translateX(2px);
-}
-
-.home-page.dark-mode .conversation-item:hover {
-  background-color: #4a4a4a;
-  border-color: var(--color-primary);
-}
-
-.conversation-name {
-  font-weight: bold;
-  color: var(--color-primary);
-  font-size: 0.9rem;
-}
-
-.conversation-last-message {
-  color: #666;
-  font-size: 0.8rem;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.home-page.dark-mode .conversation-last-message {
-  color: #aaa;
-}
-
-.conversation-time {
-  color: #999;
-  font-size: 0.7rem;
-}
-
-.home-page.dark-mode .conversation-time {
-  color: #777;
-}
-
-.notification-item {
-  padding: 0.75rem;
-  background-color: white;
-  border: 2px solid var(--color-secondary);
-  border-radius: 4px;
-  cursor: pointer;
-  transition: background-color 0.2s ease, transform 0.2s ease;
-  display: flex;
-  flex-direction: column;
-  gap: 0.25rem;
-}
-
-.home-page.dark-mode .notification-item {
-  background-color: #333;
-  border-color: var(--color-secondary);
-}
-
-.notification-item:hover {
-  background-color: #f9f9f9;
-  transform: translateX(2px);
-}
-
-.home-page.dark-mode .notification-item:hover {
-  background-color: #3a3a3a;
-}
-
-.notification-sender {
-  font-weight: bold;
-  color: var(--color-secondary);
-  font-size: 0.85rem;
-}
-
-.notification-content {
-  color: #666;
-  font-size: 0.8rem;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.home-page.dark-mode .notification-content {
-  color: #aaa;
-}
-
-.notification-time {
-  color: #999;
-  font-size: 0.7rem;
-}
-
-.home-page.dark-mode .notification-time {
-  color: #777;
 }
