--- frontend/src/pages/home/itemPostForm/itemPostForm.css
@@ -1,309 +1,0 @@
-/* Modal overlay */
-.item-post-form-overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background-color: rgba(0, 0, 0, 0.6);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 1000;
-  animation: fadeIn 0.3s ease-out;
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-
-/* Modal container */
-.item-post-form-container {
-  background-color: var(--color-background-light);
-  border-radius: 12px;
-  padding: 2rem;
-  max-width: 500px;
-  width: 90%;
-  max-height: 90vh;
-  overflow-y: auto;
-  box-shadow: 0 10px 40px rgba(0, 0, 0, 0.3);
-  animation: slideUp 0.3s ease-out;
-  position: relative;
-}
-
-.item-post-form-overlay.dark-mode .item-post-form-container {
-  background-color: var(--color-background-dark);
-}
-
-@keyframes slideUp {
-  from {
-    transform: translateY(30px);
-    opacity: 0;
-  }
-  to {
-    transform: translateY(0);
-    opacity: 1;
-  }
-}
-
-/* Close button */
-.item-post-form-close {
-  position: absolute;
-  top: 1rem;
-  right: 1rem;
-  background: none;
-  border: none;
-  font-size: 2rem;
-  cursor: pointer;
-  color: var(--color-primary);
-  padding: 0;
-  width: 2.5rem;
-  height: 2.5rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: transform 0.2s ease;
-}
-
-.item-post-form-close:hover {
-  transform: scale(1.2);
-}
-
-/* Title */
-.item-post-form-title {
-  margin-top: 0;
-  margin-bottom: 1.5rem;
-  font-size: 1.8rem;
-  font-weight: 700;
-  color: black;
-}
-
-.item-post-form-overlay.dark-mode .item-post-form-title {
-  color: white;
-}
-
-/* Form */
-.item-post-form {
-  display: flex;
-  flex-direction: column;
-  gap: 1.2rem;
-}
-
-/* Form group */
-.item-post-form-group {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.item-post-form-group label {
-  font-size: 0.95rem;
-  font-weight: 600;
-  color: black;
-}
-
-.item-post-form-overlay.dark-mode .item-post-form-group label {
-  color: white;
-}
-
-/* Form inputs */
-.item-post-form-input,
-.item-post-form-textarea {
-  padding: 0.75rem 1rem;
-  border: 2px solid var(--color-primary);
-  border-radius: 6px;
-  font-size: 0.95rem;
-  font-family: inherit;
-  background-color: var(--color-background-light);
-  color: black;
-  outline: none;
-  transition: border-color 0.3s ease, box-shadow 0.3s ease;
-}
-
-.item-post-form-overlay.dark-mode .item-post-form-input,
-.item-post-form-overlay.dark-mode .item-post-form-textarea {
-  background-color: #444444;
-  color: white;
-}
-
-.item-post-form-input:focus,
-.item-post-form-textarea:focus {
-  border-color: var(--color-secondary);
-  box-shadow: 0 0 0 3px rgba(173, 139, 115, 0.1);
-}
-
-/* Textarea */
-.item-post-form-textarea {
-  resize: vertical;
-  min-height: 100px;
-}
-
-/* Form buttons container */
-.item-post-form-buttons {
-  display: flex;
-  gap: 1rem;
-  margin-top: 1rem;
-}
-
-/* Submit button */
-.item-post-form-submit {
-  flex: 1;
-  padding: 0.9rem 1.5rem;
-  background-color: var(--color-secondary);
-  color: white;
-  border: none;
-  border-radius: 6px;
-  font-size: 1rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: transform 0.2s ease, box-shadow 0.2s ease;
-}
-
-.item-post-form-submit:hover:not(:disabled) {
-  transform: translateY(-2px);
-  box-shadow: 0 4px 12px rgba(250, 205, 0, 0.3);
-}
-
-.item-post-form-submit:active:not(:disabled) {
-  transform: translateY(0);
-}
-
-.item-post-form-submit:disabled {
-  opacity: 0.6;
-  cursor: not-allowed;
-}
-
-/* Cancel button */
-.item-post-form-cancel {
-  flex: 1;
-  padding: 0.9rem 1.5rem;
-  background-color: transparent;
-  color: var(--color-primary);
-  border: 2px solid var(--color-primary);
-  border-radius: 6px;
-  font-size: 1rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: background-color 0.2s ease, color 0.2s ease;
-}
-
-.item-post-form-cancel:hover:not(:disabled) {
-  background-color: var(--color-primary);
-  color: white;
-}
-
-.item-post-form-cancel:disabled {
-  opacity: 0.6;
-  cursor: not-allowed;
-}
-
-/* Success message */
-.item-post-form-success {
-  background-color: #4caf50;
-  color: white;
-  padding: 1rem;
-  border-radius: 6px;
-  margin-bottom: 1rem;
-  text-align: center;
-  font-weight: 600;
-  animation: slideDown 0.3s ease-out;
-}
-
-@keyframes slideDown {
-  from {
-    transform: translateY(-10px);
-    opacity: 0;
-  }
-  to {
-    transform: translateY(0);
-    opacity: 1;
-  }
-}
-
-/* Error message */
-.item-post-form-error {
-  background-color: #f44336;
-  color: white;
-  padding: 1rem;
-  border-radius: 6px;
-  margin-bottom: 1rem;
-  text-align: center;
-  font-weight: 600;
-  animation: slideDown 0.3s ease-out;
-}
-
-/* Modal note */
-.item-post-form-note {
-  font-size: 0.85rem;
-  color: #666;
-  margin-top: 1.5rem;
-  text-align: center;
-  font-style: italic;
-}
-
-.item-post-form-overlay.dark-mode .item-post-form-note {
-  color: #999;
-}
-
-/* Mobile responsive */
-@media (max-width: 480px) {
-  .item-post-form-container {
-    width: 95%;
-    padding: 1.5rem 1rem;
-    max-height: 95vh;
-  }
-
-  .item-post-form-title {
-    font-size: 1.4rem;
-    margin-bottom: 1rem;
-  }
-
-  .item-post-form-group {
-    gap: 0.4rem;
-  }
-
-  .item-post-form-group label {
-    font-size: 0.9rem;
-  }
-
-  .item-post-form-input,
-  .item-post-form-textarea {
-    font-size: 0.9rem;
-    padding: 0.7rem 0.9rem;
-  }
-
-  .item-post-form-textarea {
-    min-height: 80px;
-  }
-
-  .item-post-form-buttons {
-    flex-direction: column;
-    gap: 0.8rem;
-  }
-
-  .item-post-form-submit,
-  .item-post-form-cancel {
-    padding: 0.8rem 1rem;
-    font-size: 0.95rem;
-  }
-
-  .item-post-form-close {
-    width: 2rem;
-    height: 2rem;
-    font-size: 1.5rem;
-    top: 0.75rem;
-    right: 0.75rem;
-  }
-}
-
-/* Tablet responsive */
-@media (max-width: 768px) and (orientation: portrait) {
-  .item-post-form-container {
-    width: 85%;
-  }
-}
