--- frontend/src/pages/home/liveChat/liveChat.css
@@ -1,280 +1,0 @@
-.chat-container {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  background-color: #f5f5f5;
-  border-radius: 8px;
-  overflow: hidden;
-}
-
-.chat-header {
-  padding: 1rem;
-  background-color: #ff6b6b;
-  color: white;
-  font-weight: 600;
-  border-bottom: 2px solid #e63946;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.chat-header-title {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.online-indicator {
-  width: 12px;
-  height: 12px;
-  background-color: #00d084;
-  border-radius: 50%;
-  display: inline-block;
-}
-
-.messages-container {
-  flex: 1;
-  overflow-y: auto;
-  padding: 1rem;
-  display: flex;
-  flex-direction: column;
-  gap: 0.75rem;
-  background-color: #ffffff;
-}
-
-.message {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 0.5rem;
-  animation: slideIn 0.3s ease-in-out;
-}
-
-@keyframes slideIn {
-  from {
-    opacity: 0;
-    transform: translateY(10px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-.message.sent {
-  align-items: flex-end;
-}
-
-.message.received {
-  align-items: flex-start;
-}
-
-.message-bubble {
-  max-width: 70%;
-  padding: 0.75rem 1rem;
-  border-radius: 12px;
-  word-wrap: break-word;
-  word-break: break-word;
-}
-
-.message.sent .message-bubble {
-  background-color: #ff6b6b;
-  color: white;
-  border-bottom-right-radius: 4px;
-}
-
-.message.received .message-bubble {
-  background-color: #e8e8e8;
-  color: #333;
-  border-bottom-left-radius: 4px;
-}
-
-.message-sender {
-  font-size: 0.75rem;
-  color: #666;
-  margin-bottom: 0.25rem;
-  font-weight: 500;
-}
-
-.message.sent .message-sender {
-  color: #666;
-  text-align: right;
-}
-
-.message.received .message-sender {
-  color: #999;
-  text-align: left;
-}
-
-.message-content {
-  word-break: break-word;
-  overflow-wrap: break-word;
-}
-
-.message-time {
-  font-size: 0.7rem;
-  color: #999;
-  margin-top: 0.25rem;
-}
-
-.message.sent .message-time {
-  text-align: right;
-  color: #999;
-}
-
-.message.received .message-time {
-  text-align: left;
-  color: #bbb;
-}
-
-.empty-messages {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: #999;
-  font-size: 0.9rem;
-  text-align: center;
-  padding: 2rem;
-}
-
-.chat-input-container {
-  padding: 1rem;
-  background-color: #f9f9f9;
-  border-top: 1px solid #ddd;
-  display: flex;
-  gap: 0.5rem;
-  align-items: center;
-}
-
-.chat-input-field {
-  flex: 1;
-  padding: 0.75rem;
-  border: 1px solid #ddd;
-  border-radius: 6px;
-  font-size: 0.9rem;
-  font-family: inherit;
-}
-
-.chat-input-field:focus {
-  outline: none;
-  border-color: #ff6b6b;
-  box-shadow: 0 0 0 3px rgba(255, 107, 107, 0.1);
-}
-
-.chat-send-button {
-  padding: 0.75rem 1.5rem;
-  background-color: #ff6b6b;
-  color: white;
-  border: none;
-  border-radius: 6px;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.3s ease;
-}
-
-.chat-send-button:hover {
-  background-color: #e63946;
-  transform: translateY(-2px);
-  box-shadow: 0 4px 12px rgba(230, 57, 70, 0.3);
-}
-
-.chat-send-button:active {
-  transform: translateY(0);
-}
-
-.chat-send-button:disabled {
-  background-color: #ccc;
-  cursor: not-allowed;
-  transform: none;
-}
-
-.typing-indicator {
-  display: flex;
-  gap: 0.25rem;
-  padding: 0.75rem 1rem;
-  background-color: #e8e8e8;
-  border-radius: 12px;
-  width: fit-content;
-}
-
-.typing-dot {
-  width: 8px;
-  height: 8px;
-  background-color: #999;
-  border-radius: 50%;
-  animation: typing 1.4s infinite;
-}
-
-.typing-dot:nth-child(2) {
-  animation-delay: 0.2s;
-}
-
-.typing-dot:nth-child(3) {
-  animation-delay: 0.4s;
-}
-
-@keyframes typing {
-  0%, 60%, 100% {
-    opacity: 0.5;
-    transform: translateY(0);
-  }
-  30% {
-    opacity: 1;
-    transform: translateY(-10px);
-  }
-}
-
-.connection-status {
-  padding: 0.5rem 1rem;
-  background-color: #fff3cd;
-  color: #856404;
-  font-size: 0.8rem;
-  text-align: center;
-  border-bottom: 1px solid #ffc107;
-}
-
-.connection-status.connected {
-  background-color: #d4edda;
-  color: #155724;
-  border-bottom-color: #28a745;
-}
-
-.connection-status.disconnected {
-  background-color: #f8d7da;
-  color: #721c24;
-  border-bottom-color: #f5c6cb;
-}
-
-/* Responsive design */
-@media (max-width: 768px) {
-  .message-bubble {
-    max-width: 85%;
-  }
-
-  .chat-input-container {
-    padding: 0.75rem;
-  }
-
-  .chat-send-button {
-    padding: 0.6rem 1rem;
-    font-size: 0.85rem;
-  }
-}
-
-@media (max-width: 480px) {
-  .message-bubble {
-    max-width: 90%;
-  }
-
-  .chat-header {
-    padding: 0.75rem;
-  }
-
-  .messages-container {
-    padding: 0.75rem;
-  }
-
-  .chat-input-container {
-    padding: 0.5rem;
-  }
-}
