--- frontend/src/pages/myItems/myItems.css
@@ -1,277 +1,0 @@
-* {
-  box-sizing: border-box;
-}
-
-.my-items-page {
-  display: flex;
-  flex-direction: column;
-  height: 100vh;
-  width: 100%;
-  overflow-x: hidden;
-  background-color: var(--color-background-light);
-  transition: background-color 0.3s ease;
-}
-
-.my-items-page.light-mode {
-  background-color: var(--color-background-light);
-  color: black;
-}
-
-.my-items-page.dark-mode {
-  background-color: var(--color-background-dark);
-  color: #e0e0e0;
-}
-
-.my-items-column {
-  flex: 1;
-  overflow-y: auto;
-  padding: 1.5rem;
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-  width: 100%;
-  transition: background-color 0.3s ease;
-  background-color: var(--color-background-light);
-}
-
-.my-items-page.dark-mode .my-items-column {
-  background-color: var(--color-background-dark);
-}
-
-.my-items-header {
-  padding-bottom: 1rem;
-  border-bottom: 2px solid var(--color-primary);
-  margin-bottom: 1rem;
-}
-
-.my-items-header h2 {
-  margin: 0;
-  font-size: 2rem;
-  font-weight: 700;
-  color: black;
-}
-
-.my-items-page.dark-mode .my-items-header h2 {
-  color: white;
-}
-
-.my-item-row {
-  display: grid;
-  grid-template-columns: 100px 1fr 150px;
-  gap: 1rem;
-  padding: 1vw 4vh;
-  padding-right: 4vw;
-  border: 3px solid var(--color-primary);
-  border-radius: 8px;
-  background-color: var(--color-background-light);
-  align-items: start;
-  transition: background-color 0.3s ease;
-}
-
-.my-items-page.dark-mode .my-item-row {
-  background-color: #333333;
-  border-color: var(--color-primary);
-}
-
-.my-item-image {
-  font-size: 3rem;
-  text-align: center;
-  line-height: 1;
-  width: 150px;
-  height: 150px;
-  aspect-ratio: 1 / 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  overflow: hidden;
-  border-radius: 4px;
-  background-color: #f5f5f5;
-  flex-shrink: 0;
-}
-
-.my-item-image img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.my-item-details {
-  display: flex;
-  flex-direction: column;
-  gap: 1vh;
-  margin-left: 3vw;
-}
-
-.my-item-what {
-  font-size: 2rem;
-  font-weight: 700;
-  color: black;
-  margin: 0;
-}
-
-.my-items-page.dark-mode .my-item-what {
-  color: white;
-}
-
-.my-item-description {
-  font-size: 1.2rem;
-  color: black;
-  line-height: 1.4;
-  font-weight: 400;
-}
-
-.my-items-page.dark-mode .my-item-description {
-  color: white;
-}
-
-.my-item-where,
-.my-item-location,
-.my-item-when {
-  font-size: 1.1rem;
-  color: black;
-  margin: 0;
-  line-height: 1.3;
-}
-
-.my-items-page.dark-mode .my-item-where,
-.my-items-page.dark-mode .my-item-location,
-.my-items-page.dark-mode .my-item-when {
-  color: white;
-}
-
-.my-item-right {
-  display: flex;
-  flex-direction: column;
-  gap: 2vh;
-  align-items: flex-start;
-  width: 10vw;
-}
-
-.my-item-type-badge {
-  padding: 0.5rem 1rem;
-  background-color: var(--color-secondary);
-  color: white;
-  border-radius: 4px;
-  font-size: 0.9rem;
-  font-weight: 600;
-}
-
-.my-item-delete-button {
-  padding: 1.8vh 1.6vw;
-  background-color: #f44336;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  font-weight: 600;
-  font-size: 1.2rem;
-  cursor: pointer;
-  transition: transform 0.2s ease, box-shadow 0.2s ease, background-color 0.2s ease;
-  white-space: nowrap;
-}
-
-.my-item-delete-button:hover:not(:disabled) {
-  background-color: #d32f2f;
-  transform: translateY(-1.5px);
-  box-shadow: 0 2px 6px rgba(244, 67, 54, 0.3);
-}
-
-.my-item-delete-button:active:not(:disabled) {
-  transform: translateY(0px);
-  box-shadow: 0 1px 3px rgba(244, 67, 54, 0.2);
-}
-
-.my-item-delete-button:disabled {
-  opacity: 0.6;
-  cursor: not-allowed;
-}
-
-/* Tablet */
-@media (max-width: 820px) and (orientation: portrait) {
-  .my-item-row {
-    grid-template-columns: 80px 1fr 120px;
-    gap: 0.75rem;
-    padding: 1rem;
-  }
-
-  .my-item-image {
-    font-size: 2.5rem;
-    min-width: 60px;
-    width: 15vh;
-    height: 15vh;
-  }
-
-  .my-item-delete-button {
-    font-size: 0.8rem;
-    padding: 0.5rem 0.8rem;
-  }
-
-  .my-item-right {
-    width: 15vw;
-  }
-
-  .my-item-details {
-    margin-left: 12vw;
-  }
-
-  .my-item-what {
-    font-size: 1.7rem;
-  }
-
-  .my-item-where,
-  .my-item-location,
-  .my-item-when {
-    font-size: 1.1rem;
-  }
-}
-
-/* Mobile */
-@media (max-width: 480px) {
-  .my-items-column {
-    padding: 1rem;
-  }
-
-  .my-item-row {
-    grid-template-columns: 1fr;
-    flex-direction: column;
-    gap: 0.5rem;
-    padding: 0.75rem;
-    border: 1px solid var(--color-primary);
-    display: flex;
-  }
-
-  .my-item-image {
-    font-size: 2rem;
-    min-width: 50px;
-    width: 16vh;
-    height: 16vh;
-  }
-
-  .my-item-details {
-    margin-left: 0vw;
-    margin-top: 2vh;
-  }
-
-  .my-item-right {
-    width: 100%;
-    flex-direction: row;
-    gap: 1rem;
-    align-items: center;
-    margin-top: 1rem;
-    padding-top: 1rem;
-    border-top: 1px solid var(--color-primary);
-  }
-
-  .my-item-delete-button {
-    flex: 1;
-    padding: 0.8rem 1rem;
-    font-size: 0.9rem;
-  }
-
-  .my-item-what {
-    font-size: 1.8rem;
-  }
-
-  .my-items-header h2 {
-    font-size: 1.4rem;
-  }
-}
