--- frontend/src/pages/myItems/myItems.tsx
@@ -1,218 +1,0 @@
-import { useState, useEffect } from 'react'
-import { useNavigate } from 'react-router-dom'
-import './myItems.css'
-import Navbar from '../navbar/navbar'
-
-interface ItemImage {
-  url: string
-  filename: string
-  uploadedAt: string
-}
-
-interface ItemLocation {
-  type: string
-  coordinates: number[]
-  buildingName: string
-}
-
-interface DBItem {
-  _id: string
-  userId: { _id: string; name: string; email: string } | string
-  type: 'lost' | 'found'
-  title: string
-  description: string
-  location: ItemLocation
-  images: ItemImage[]
-  tags: string[]
-  status: string
-  matchCount: number
-  createdAt: string
-  updatedAt: string
-}
-
-interface UserItem extends DBItem {
-  id: string
-  image: string
-  what: string
-  where: string
-  location_display: string
-  when: string
-  founder: string
-}
-
-const API_BASE_URL = 'http://localhost:5000/api'
-
-export default function MyItems() {
-  const [items, setItems] = useState<UserItem[]>([])
-  const [isDarkMode, setIsDarkMode] = useState(false)
-  const [loading, setLoading] = useState(true)
-  const [error, setError] = useState<string | null>(null)
-  const [deletingId, setDeletingId] = useState<string | null>(null)
-  const navigate = useNavigate()
-
-  useEffect(() => {
-    fetchMyItems()
-  }, [])
-
-  const handleThemeToggle = (isDark: boolean) => {
-    setIsDarkMode(isDark)
-    localStorage.setItem('theme', isDark ? 'dark' : 'light')
-  }
-
-  const handleLogout = () => {
-    localStorage.removeItem('authToken')
-    localStorage.removeItem('user')
-    window.location.href = '/login'
-  }
-
-  const handleLogoClick = () => {
-    navigate('/')
-  }
-
-  const handlePostClick = () => {
-    // Navigate back to home
-    navigate('/')
-  }
-
-  const fetchMyItems = async () => {
-    try {
-      setLoading(true)
-      setError(null)
-
-      const authToken = localStorage.getItem('authToken')
-      if (!authToken) {
-        setError('Authentication required')
-        return
-      }
-
-      const response = await fetch(`${API_BASE_URL}/items/my`, {
-        headers: {
-          'Authorization': `Bearer ${authToken}`
-        }
-      })
-
-      if (!response.ok) {
-        throw new Error('Failed to fetch your items')
-      }
-
-      const data = await response.json()
-
-      // Transform backend data to frontend format
-      const transformedItems: UserItem[] = data.items.map((item: DBItem) => ({
-        ...item,
-        id: item._id,
-        image: item.images.length > 0 ? item.images[0].url : '',
-        what: item.title,
-        where: `Wo wurde gefunden: ${item.location.buildingName}`,
-        location_display: `Wo zu finden ist: ${item.location.buildingName}`,
-        when: `Wann wurde gefunden: ${new Date(item.createdAt).toLocaleDateString('de-DE')} ${new Date(item.createdAt).toLocaleTimeString('de-DE', { hour: '2-digit', minute: '2-digit' })}`,
-        founder: typeof item.userId === 'string' ? 'Unknown' : item.userId.name || 'Unknown'
-      }))
-
-      setItems(transformedItems)
-    } catch (err) {
-      const message = err instanceof Error ? err.message : 'An error occurred'
-      setError(message)
-      console.error('Error fetching items:', err)
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  const handleDeleteItem = async (itemId: string) => {
-    if (!window.confirm('Are you sure you want to delete this item?')) {
-      return
-    }
-
-    try {
-      setDeletingId(itemId)
-
-      const authToken = localStorage.getItem('authToken')
-      if (!authToken) {
-        setError('Authentication required')
-        return
-      }
-
-      const response = await fetch(`${API_BASE_URL}/items/${itemId}`, {
-        method: 'DELETE',
-        headers: {
-          'Authorization': `Bearer ${authToken}`
-        }
-      })
-
-      if (!response.ok) {
-        throw new Error('Failed to delete item')
-      }
-
-      // Remove item from state
-      setItems(items.filter(item => item.id !== itemId))
-    } catch (err) {
-      const message = err instanceof Error ? err.message : 'An error occurred'
-      setError(message)
-      console.error('Error deleting item:', err)
-    } finally {
-      setDeletingId(null)
-    }
-  }
-
-  return (
-    <div className={`my-items-page ${isDarkMode ? 'dark-mode' : 'light-mode'}`}>
-      <Navbar 
-        onMessageClick={() => {}}
-        onLogout={handleLogout}
-        onLogoClick={handleLogoClick}
-        isDarkMode={isDarkMode}
-        onThemeToggle={handleThemeToggle}
-        onPostClick={handlePostClick}
-      />
-
-      {/* Main content */}
-      <div className="my-items-column">
-        <div className="my-items-header">
-          <h2>Meine Posts</h2>
-        </div>
-
-        {loading ? (
-          <p style={{ textAlign: 'center', padding: '2rem' }}>Loading your items...</p>
-        ) : error ? (
-          <p style={{ textAlign: 'center', padding: '2rem', color: 'red' }}>Error: {error}</p>
-        ) : items.length === 0 ? (
-          <p style={{ textAlign: 'center', padding: '2rem' }}>You haven't posted any items yet</p>
-        ) : (
-          items.map((item) => (
-            <div key={item.id} className="my-item-row">
-              <div className="my-item-image">
-                {item.image ? (
-                  <img src={item.image} alt={item.what} />
-                ) : (
-                  <div style={{ width: '100%', height: '100%', display: 'flex', alignItems: 'center', justifyContent: 'center', backgroundColor: '#e0e0e0' }}>
-                    No Image
-                  </div>
-                )}
-              </div>
-
-              <div className="my-item-details">
-                <h3 className="my-item-what">{item.what}</h3>
-                <p className="my-item-description">Description: {item.description}</p>
-                <p className="my-item-where">{item.where}</p>
-                <p className="my-item-location">{item.location_display}</p>
-                <p className="my-item-when">{item.when}</p>
-              </div>
-
-              <div className="my-item-right">
-                <span className="my-item-type-badge">{item.type.toUpperCase()}</span>
-                <button
-                  className="my-item-delete-button"
-                  onClick={() => handleDeleteItem(item.id)}
-                  disabled={deletingId === item.id}
-                >
-                  {deletingId === item.id ? 'Deleting...' : 'Löschen'}
-                </button>
-              </div>
-            </div>
-          ))
-        )}
-      </div>
-    </div>
-  )
-}
